Relocates class CollectingObserver to grpc-core.
